+++ src/main/java/com/tiketeer/Tiketeer/domain/purchase/controller/PurchaseController.java
@@ -15,7 +15,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.tiketeer.Tiketeer.domain.purchase.controller.dto.DeletePurchaseTicketsRequestDto;
 import com.tiketeer.Tiketeer.domain.purchase.controller.dto.PostPurchaseRequestDto;
 import com.tiketeer.Tiketeer.domain.purchase.controller.dto.PostPurchaseResponseDto;
-import com.tiketeer.Tiketeer.domain.purchase.service.PurchaseService;
+import com.tiketeer.Tiketeer.domain.purchase.usecase.CreatePurchaseUseCase;
+import com.tiketeer.Tiketeer.domain.purchase.usecase.DeletePurchaseTicketsUseCase;
 import com.tiketeer.Tiketeer.response.ApiResponse;
 
 import jakarta.validation.Valid;
@@ -23,18 +24,21 @@ import jakarta.validation.Valid;
 @RestController
 @RequestMapping("/purchases")
 public class PurchaseController {
-	private final PurchaseService purchaseService;
+	private final CreatePurchaseUseCase createPurchaseUseCase;
+	private final DeletePurchaseTicketsUseCase deletePurchaseTicketsUseCase;
 
 	@Autowired
-	PurchaseController(PurchaseService purchaseService) {
-		this.purchaseService = purchaseService;
+	PurchaseController(CreatePurchaseUseCase createPurchaseUseCase,
+		DeletePurchaseTicketsUseCase deletePurchaseTicketsUseCase) {
+		this.createPurchaseUseCase = createPurchaseUseCase;
+		this.deletePurchaseTicketsUseCase = deletePurchaseTicketsUseCase;
 	}
 
 	@PostMapping("/")
 	public ResponseEntity<ApiResponse<PostPurchaseResponseDto>> postPurchase(
 		@Valid @RequestBody PostPurchaseRequestDto request) {
 		var memberEmail = "mock@mock.com";
-		var result = purchaseService.createPurchase(request.convertToDto(memberEmail));
+		var result = createPurchaseUseCase.createPurchase(request.convertToDto(memberEmail));
 		var responseBody = ApiResponse.wrap(PostPurchaseResponseDto.converFromDto(result));
 		return ResponseEntity.status(HttpStatus.CREATED).body(responseBody);
 	}
@@ -43,7 +47,7 @@ public class PurchaseController {
 	public ResponseEntity deletePurchaseTickets(@PathVariable UUID purchaseId, @Valid @RequestBody
 	DeletePurchaseTicketsRequestDto request) {
 		var memberEmail = "mock@mock.com";
-		purchaseService.deletePurchaseTickets(request.convertToDto(memberEmail, purchaseId));
+		deletePurchaseTicketsUseCase.deletePurchaseTickets(request.convertToDto(memberEmail, purchaseId));
 		return ResponseEntity.status(HttpStatus.OK).build();
 	}
 }
